requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
sqlite3
